TMOShan10: solving linear system using Eigen
Solving the linear system is reimplemented for using the Eigen library. The Eigen library offers more options for the Sparse matrix and optimizes the processing. The implementation of the OpenCV resize function is replaced by the implementation of the original MATLAB imresize function written in C language. In this commit, the bugs are fixed too.