TMOXia18: add SIGSEGV handling
This commit adds the solution for handling the SIGSEGV. If the SIGSEGV
is raised, the program exits with the exit code 1. The information
text is added to inform the user about the unexpected behavior that can
in rare situations occur. This print cannot be added to the SIGSEGV
signal handling due to it is not safe to have IO operations in the
signal handling function.